VERY basic functionality implemented. Will run
At this point, the Add File, Remove Selection, Browse, and Run buttons
are all "wired" to the controller and will act on the FileSet
accordingly. However, it's not programmed "correctly" right now - just
cobbled together.

Setting the destination with Browse needs to call the controller to set
the destination in the FileSet. The controller will need the destination
and the backup name. From there, the controller needs to append a
trailing slash and the backup name to the destination and then set that
as the destination on the FileSet.