FIX: decouple UIViewController from FileOps
Detail: FileOps was modified to use specific Swing methods to publish
progress. This created coupling between the two classes, which would
cause FileOps to crash if it was attempted to run from a non-Swing
"parent" (like the Schedule running if/when we create it).

I removed the setProgress() method that was added to FileOps, revised
the Progress, FileSet, and handleProgress methods and several other
related tweaks to enable the handleProgress() and handleCompletion
methods to work as designed, and eliminate the need to use setProgress.