FIX: copying not working on Windows
in FileOps lines 94-96 determine if the source path starts with
something like c: (on Mac it starts with \). The goal is to strip off
the root of the source. On Mac, that involves removing only the first
character, but on Windows, we need to remove the first 3 characters in
order for .resolve() to work correctly.