FIX: move try/catch block from Application to UIViewController
when saveDefaultFileSet is called, the try/catch block needs to be in
the UI so that a dialog can be presented to the user if there is an
error saving the default file set. The user won't really be able to "do"
anything about it, but we at least need to inform them.